move shuffle and repeat serialization tests to kernel tests
