+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -413,23 +413,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "shuffle_and_repeat_dataset_serialization_test",
-    size = "medium",
-    srcs = ["shuffle_and_repeat_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/experimental/ops:shuffle_ops",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-    ],
-)
-
-tf_py_test(
     name = "shuffle_dataset_serialization_test",
     size = "medium",
     srcs = ["shuffle_dataset_serialization_test.py"],
