+++ tensorflow/python/data/experimental/kernel_tests/shuffle_and_repeat_test.py
@@ -21,6 +21,7 @@ from absl.testing import parameterized
 import numpy as np
 
 from tensorflow.python.data.experimental.ops import shuffle_ops
+from tensorflow.python.data.kernel_tests import checkpoint_test_base
 from tensorflow.python.data.kernel_tests import test_base
 from tensorflow.python.data.ops import dataset_ops
 from tensorflow.python.framework import combinations
@@ -150,6 +151,18 @@ class ShuffleAndRepeatTest(test_base.DatasetTestBase, parameterized.TestCase):
 
     self.assertCountEqual(shuffle_1, shuffle_2)
     self.assertNotEqual(shuffle_1, shuffle_2)
+
+
+class ShuffleAndRepeatCheckpointTest(checkpoint_test_base.CheckpointTestBase,
+                                     parameterized.TestCase):
+
+  def _build_ds(self, seed):
+    return dataset_ops.Dataset.range(20).apply(
+        shuffle_ops.shuffle_and_repeat(buffer_size=5, count=5, seed=seed))
+
+  @combinations.generate(test_base.default_test_combinations())
+  def testCore(self):
+    self.run_core_tests(lambda: self._build_ds(10), 100)
 
 
 if __name__ == "__main__":
